+++ CHANGELOG.rst
@@ -10,6 +10,10 @@ This document records the main changes to the tree code.
 ------------------
 - Updating tree config for dr17 VACs
 - Added new Sphinx doc page for computing changes in the tree
+- Adds ``release_date`` to tree configs for DR and MPLs.  Unreleased configs have ``release_date`` set to None.
+- Adds new ``release_date`` property to ``Tree`` which returns the release date as a ``datetime.date`` object.
+- Expanding ``Tree.list_available_configs`` to sort by "DR", "MPL" and "IPL".
+- Deprecating ``Tree.list_configs`` in favor of ``Tree.list_available_configs``
 
 3.0.7 (2020-03-05)
 ------------------